+++ ThreeThirty_FE/threethirty/screens/CommunityScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const CommunityScreen = () => {
+  return <Text>Community</Text>;
+};
+
+export default CommunityScreen;
