+++ ThreeThirty_FE/threethirty/screens/FollowingScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const FollowingScreen = () => {
+  return <Text>Following</Text>;
+};
+
+export default FollowingScreen;
