+++ ThreeThirty_FE/threethirty/screens/FunnyScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const FunnyScreen = () => {
+  return <Text>Funny</Text>;
+};
+
+export default FunnyScreen;
