+++ ThreeThirty_FE/threethirty/screens/HomeScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const HomeScreen = () => {
+  return <Text>Home</Text>;
+};
+
+export default HomeScreen;
