+++ ThreeThirty_FE/threethirty/screens/NotificationScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const NotificationScreen = () => {
+  return <Text>Notification</Text>;
+};
+
+export default NotificationScreen;
