+++ ThreeThirty_FE/threethirty/screens/SearchScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const SearchScreen = () => {
+  return <Text>Search</Text>;
+};
+
+export default SearchScreen;
