+++ ThreeThirty_FE/threethirty/screens/ThreeThirtyScreen.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import {Text} from 'react-native';
+
+const ThreeThirtyScreen = () => {
+  return <Text>ThreeThirty</Text>;
+};
+
+export default ThreeThirtyScreen;
